chore: Add FindFakeIt.cmake (#11)
FakeIt (https://github.com/eranpeer/FakeIt) is a header-only c++ mocking
library that is very useful in tests in general and in our firmware
architectures, which have sections that cannot be compiled outside of a
microcontroller environment, in particular. This PR adds a cmake module
to find and provide FakeIt.