feat: add include-what-you-use
Adds the include-what-you-use tool. This is another lint step that maeks
sure only appropriate headers are present, and that all required headers
are present (even if things are already compiling because those headers
are in some transitive #include). This makes things faster to compile
and also safer since you can't remove an include that doesn't seem like
it has anything to do with anything and suddenly break stuff.